feat: orbslam2_driver up to STEREO mode, build fixes
@fabrizioromanelli
This is the first version of a fully-fledged ORB-SLAM2 driver node for ROS 2.
It aims at supporting many features, but currently only works in STEREO mode.
It tracks the position of both the system and the base_link on which it is mounted.
Planned short term upgrades inclued:
  - Covariance estimation.
  - FrameViewer support (now only supports the original GUI).
  - Full+ROI map pointcloud publishing.
  - Understanding why the **** I have to link against all of the other ORB-SLAM2 dependencies explicitly (ament should make me avoid having to do this).
Does it also work? We'll find out soon.